Organize StreamdeckContext into ExportMonitors and SendActions folders
