+++ Sources/StreamdeckContext/ExportMonitors/ComparisonMonitor.cpp
