+++ Sources/StreamdeckContext/ExportMonitors/ComparisonMonitor.h
