+++ Sources/StreamdeckContext/SendActions/IncrementContext.cpp
@@ -4,7 +4,7 @@
 
 #include "IncrementContext.h"
 
-#include "../Common/EPLJSONUtils.h"
+#include "../../Common/EPLJSONUtils.h"
 
 void IncrementContext::handleButtonEvent(DcsInterface &dcs_interface, const KeyEvent event, const json &inPayload)
 {
