+++ Sources/StreamdeckContext/SendActions/IncrementContext.h
@@ -2,12 +2,12 @@
 
 #pragma once
 
-#include "StreamdeckContext.h"
+#include "../StreamdeckContext.h"
 
 #include <optional>
 #include <string>
 
-class MomentaryContext : public StreamdeckContext
+class IncrementContext : public StreamdeckContext
 {
   public:
     using StreamdeckContext::StreamdeckContext;
@@ -17,7 +17,6 @@ class MomentaryContext : public StreamdeckContext
      *
      * @param dcs_interface Interface to DCS containing current game state.
      * @param event Type of button event - PRESSED or RELEASED.
-     * @param action Type of button action - used to determine momentary, swtich, or increment button type.
      * @param payload Json payload received with KeyDown/KeyUp callback.
      */
     void handleButtonEvent(DcsInterface &dcs_interface, const KeyEvent event, const json &inPayload);
@@ -30,5 +29,5 @@ class MomentaryContext : public StreamdeckContext
      * @param settings Settings for context.
      * @return (Optional) Value to be sent to Button ID if it exists.
      */
-    std::optional<std::string> determineSendValue(const KeyEvent event, const json &settings) const;
+    std::optional<std::string> determineSendValue(const KeyEvent event, const json &settings);
 };
