+++ Sources/StreamdeckContext/SendActions/MomentaryContext.h
@@ -2,12 +2,12 @@
 
 #pragma once
 
-#include "StreamdeckContext.h"
+#include "../StreamdeckContext.h"
 
 #include <optional>
 #include <string>
 
-class IncrementContext : public StreamdeckContext
+class MomentaryContext : public StreamdeckContext
 {
   public:
     using StreamdeckContext::StreamdeckContext;
@@ -17,6 +17,7 @@ class IncrementContext : public StreamdeckContext
      *
      * @param dcs_interface Interface to DCS containing current game state.
      * @param event Type of button event - PRESSED or RELEASED.
+     * @param action Type of button action - used to determine momentary, swtich, or increment button type.
      * @param payload Json payload received with KeyDown/KeyUp callback.
      */
     void handleButtonEvent(DcsInterface &dcs_interface, const KeyEvent event, const json &inPayload);
@@ -29,5 +30,5 @@ class IncrementContext : public StreamdeckContext
      * @param settings Settings for context.
      * @return (Optional) Value to be sent to Button ID if it exists.
      */
-    std::optional<std::string> determineSendValue(const KeyEvent event, const json &settings);
+    std::optional<std::string> determineSendValue(const KeyEvent event, const json &settings) const;
 };
