+++ Sources/StreamdeckContext/SendActions/SwitchContext.cpp
@@ -4,7 +4,7 @@
 
 #include "SwitchContext.h"
 
-#include "../Common/EPLJSONUtils.h"
+#include "../../Common/EPLJSONUtils.h"
 
 void SwitchContext::handleButtonEvent(DcsInterface &dcs_interface, const KeyEvent event, const json &inPayload)
 {
