+++ Sources/Windows/com.ctytler.dcs.sdPlugin.vcxproj
@@ -189,14 +189,14 @@
     <ClInclude Include="..\DcsInterface\DcsIdLookup.h" />
     <ClInclude Include="..\DcsInterface\DcsInterface.h" />
     <ClInclude Include="..\DcsInterface\DcsInterfaceParameters.h" />
-    <ClInclude Include="..\StreamdeckContext\IncrementContext.h" />
-    <ClInclude Include="..\StreamdeckContext\MomentaryContext.h" />
-    <ClInclude Include="..\StreamdeckContext\Monitors\ComparisonMonitor.h" />
-    <ClInclude Include="..\StreamdeckContext\Monitors\IncrementMonitor.h" />
-    <ClInclude Include="..\StreamdeckContext\Monitors\TitleMonitor.h" />
+    <ClInclude Include="..\StreamdeckContext\SendActions\IncrementContext.h" />
+    <ClInclude Include="..\StreamdeckContext\SendActions\MomentaryContext.h" />
+    <ClInclude Include="..\StreamdeckContext\ExportMonitors\ComparisonMonitor.h" />
+    <ClInclude Include="..\StreamdeckContext\ExportMonitors\IncrementMonitor.h" />
+    <ClInclude Include="..\StreamdeckContext\ExportMonitors\TitleMonitor.h" />
     <ClInclude Include="..\StreamdeckContext\StreamdeckContext.h" />
     <ClInclude Include="..\StreamdeckContext\StreamdeckContextFactory.h" />
-    <ClInclude Include="..\StreamdeckContext\SwitchContext.h" />
+    <ClInclude Include="..\StreamdeckContext\SendActions\SwitchContext.h" />
     <ClInclude Include="..\Utilities\Decimal.h" />
     <ClInclude Include="..\Utilities\StringUtilities.h" />
     <ClInclude Include="..\Utilities\UdpSocket.h" />
@@ -230,14 +230,14 @@
     </ClCompile>
     <ClCompile Include="..\DcsInterface\DcsIdLookup.cpp" />
     <ClCompile Include="..\DcsInterface\DcsInterface.cpp" />
-    <ClCompile Include="..\StreamdeckContext\IncrementContext.cpp" />
-    <ClCompile Include="..\StreamdeckContext\MomentaryContext.cpp" />
-    <ClCompile Include="..\StreamdeckContext\Monitors\ComparisonMonitor.cpp" />
-    <ClCompile Include="..\StreamdeckContext\Monitors\IncrementMonitor.cpp" />
-    <ClCompile Include="..\StreamdeckContext\Monitors\TitleMonitor.cpp" />
+    <ClCompile Include="..\StreamdeckContext\SendActions\IncrementContext.cpp" />
+    <ClCompile Include="..\StreamdeckContext\SendActions\MomentaryContext.cpp" />
+    <ClCompile Include="..\StreamdeckContext\ExportMonitors\ComparisonMonitor.cpp" />
+    <ClCompile Include="..\StreamdeckContext\ExportMonitors\IncrementMonitor.cpp" />
+    <ClCompile Include="..\StreamdeckContext\ExportMonitors\TitleMonitor.cpp" />
     <ClCompile Include="..\StreamdeckContext\StreamdeckContext.cpp" />
     <ClCompile Include="..\StreamdeckContext\StreamdeckContextFactory.cpp" />
-    <ClCompile Include="..\StreamdeckContext\SwitchContext.cpp" />
+    <ClCompile Include="..\StreamdeckContext\SendActions\SwitchContext.cpp" />
     <ClCompile Include="..\Utilities\Decimal.cpp" />
     <ClCompile Include="..\Utilities\StringUtilities.cpp" />
     <ClCompile Include="..\Utilities\UdpSocket.cpp" />
